module_17/homework_17_4/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Vladimir Kolesnikov <vk.coworking@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
pydantic = "^2.9.2"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
python-slugify = "^8.0.4"
uvicorn = "^0.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
